+++ client/buildingblock_definition_version_test.go
@@ -10,7 +10,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/meshcloud/terraform-provider-meshstack/client/types"
-	"github.com/meshcloud/terraform-provider-meshstack/client/types/ptr"
 )
 
 var (
@@ -29,7 +28,7 @@ func TestMeshBuildingBlockDefinitionInput_UnmarshalJSON(t *testing.T) {
 		{"empty", false, types.SecretOrAny{}, types.SecretOrAny{}, assert.NoError},
 		{"not_sensitive", false, types.SecretOrAny{Y: true}, types.SecretOrAny{Y: "some-string"}, assert.NoError},
 		{"not_sensitive_but_hash", false, types.SecretOrAny{Y: map[string]any{"hash": "some-hash-looks-like-secret"}}, types.SecretOrAny{}, assert.NoError},
-		{"sensitive", true, types.SecretOrAny{}, types.SecretOrAny{X: types.Secret{Hash: ptr.To("some-hash")}}, assert.NoError},
+		{"sensitive", true, types.SecretOrAny{}, types.SecretOrAny{X: types.Secret{Hash: new("some-hash")}}, assert.NoError},
 		{"sensitive_but_no_hash", true, types.SecretOrAny{Y: map[string]any{}}, types.SecretOrAny{}, func(t assert.TestingT, err error, msgAndArgs ...any) bool {
 			return assert.ErrorContains(t, err, "got sensitive argument or default_value but variant Y is set instead")
 		}},